+++ examples/export.py
@@ -1,5 +1,4 @@
 # runs a sforce SOQL query and saves the results as a csv file.
-
 import sys
 import string
 import beatbox
@@ -19,16 +18,9 @@ def buildSoql(sobjectName):
 
 
 def printColumnHeaders(queryResult):
-    needsComma = 0
     # note that we offset 2 into the records child collection
     # to skip the type and base sObject id elements
-    for col in queryResult[sf.records][2:]:
-        if needsComma:
-            print ',',
-        else:
-            needsComma = 1
-        print col._name[1],
-    print
+    print(', '.join(col._name[1] for col in queryResult[sf.records][2:]))
 
 
 def export(username, password, objectOrSoql):
@@ -45,14 +37,7 @@ def export(username, password, objectOrSoql):
             printColumnHeaders(qr)
             printHeaders = 0
         for row in qr[sf.records:]:
-            needsComma = False
-            for col in row[2:]:
-                if needsComma:
-                    print ',',
-                else:
-                    needsComma = True
-                print str(col),
-            print
+            print(', '.join(str(col) for col in row[2:]))
         if str(qr[sf.done]) == 'true':
             break
         qr = svc.queryMore(str(qr[sf.queryLocator]))
@@ -60,6 +45,6 @@ def export(username, password, objectOrSoql):
 
 if __name__ == "__main__":
     if len(sys.argv) != 4:
-        print "usage is export.py <username> <password> [<sobjectName> || <soqlQuery>]"
+        print("usage is export.py <username> <password> [<sobjectName> || <soqlQuery>]")
     else:
         export(sys.argv[1], sys.argv[2], sys.argv[3])