+++ examples/soql2atom.py
@@ -63,11 +63,11 @@ def soql2atom(loginResult, soql, title):
     atom_ns = "http://www.w3.org/2005/Atom"
     ent_ns = "urn:sobject.enterprise.soap.sforce.com"
 
-    print "content-type: application/atom+xml"
-    doGzip = os.environ.has_key("HTTP_ACCEPT_ENCODING") and "gzip" in string.lower(os.environ["HTTP_ACCEPT_ENCODING"]).split(',')
+    print("content-type: application/atom+xml")
+    doGzip = "HTTP_ACCEPT_ENCODING" in os.environ and "gzip" in string.lower(os.environ["HTTP_ACCEPT_ENCODING"]).split(',')
     if doGzip:
-        print "content-encoding: gzip"
-    print ""
+        print("content-encoding: gzip")
+    print()
     x = beatbox.XmlWriter(doGzip)
     x.startPrefixMapping("a", atom_ns)
     x.startPrefixMapping("s", ent_ns)
@@ -108,7 +108,7 @@ def soql2atom(loginResult, soql, title):
         x.characters("\n")
         x.endElement()  # entry
     x.endElement()  # feed
-    print x.endDocument()
+    print(x.endDocument())
 
 
 def writeLink(x, namespace, localname, rel, type, href):
@@ -120,29 +120,29 @@ def writeLink(x, namespace, localname, rel, type, href):
 
 
 def authenticationRequired(message="Unauthorized"):
-    print "status: 401 Unauthorized"
-    print "WWW-authenticate: Basic realm=""www.salesforce.com"""
-    print "content-type: text/plain"
-    print ""
-    print message
+    print("status: 401 Unauthorized")
+    print("WWW-authenticate: Basic realm=""www.salesforce.com""")
+    print("content-type: text/plain")
+    print("")
+    print(message)
 
 
-if not os.environ.has_key('X_HTTP_AUTHORIZATION') or os.environ['X_HTTP_AUTHORIZATION'] == '':
+if 'X_HTTP_AUTHORIZATION' not in os.environ or os.environ['X_HTTP_AUTHORIZATION'] == '':
     authenticationRequired()
 else:
     auth = os.environ['X_HTTP_AUTHORIZATION']
     (username, password) = base64.decodestring(auth.split(" ")[1]).split(':')
     form = cgi.FieldStorage()
-    if not form.has_key('soql'):
+    if 'soql' not in form:
         raise Exception("Must provide the SOQL query to run via the soql queryString parameter")
     soql = form.getvalue("soql")
     title = "SOQL2ATOM : " + soql
-    if form.has_key("title"):
+    if "title" in form:
         title = form.getvalue("title")
     try:
         lr = svc.login(username, password)
         soql2atom(lr, soql, title)
-    except beatbox.SoapFaultError, sfe:
+    except beatbox.SoapFaultError as sfe:
         if (sfe.faultCode == 'INVALID_LOGIN'):
             authenticationRequired(sfe.faultString)
         else: