+++ run_tests.sh
@@ -2,5 +2,6 @@
 PYTHON=python
 export PYTHONPATH=.:./src
 
-$PYTHON src/beatbox/tests/test_beatbox.py
+# if the first test fails then the whole script must fail 
+$PYTHON src/beatbox/tests/test_beatbox.py &&
 $PYTHON src/beatbox/tests/test_pythonClient.py